# swagger.yaml
openapi: 3.0.0
info:
  title: E-commerce API
  contact: {}
  version: "1.0"
servers:
  - url: http://localhost:3000/api/v1
    variables: {}
paths:
...
  /product/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
        description: The product ID
    get:
      tags:
        - Product
      summary: Get specific product
      operationId: Getspecificproduct
      parameters: []
      responses:
        "200":
          description: "Successful fetched from database."
          headers: {}
        "400":
          description: "Product name required."
          headers: {}
        "404":
          description: "Product not found."
          headers: {}
      deprecated: false
    delete:
      tags:
        - Product
      summary: Delete Product
      operationId: DeleteProduct
      parameters: []
      responses:
        "200":
          description: "Product successfully removed from the database."
          headers: {}
        "400":
          description: "Invalid token or product does not exist."
        "401":
          description: "No token."
        "403":
          description: "You don't have the permssions to access this route."
      deprecated: false
      security:
        - bearer: []
    patch:
      tags:
        - Product
      summary: Update product
      operationId: Updateproduct
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/UpdateproductRequest"
                - example:
                    name: boby
                    price: 456.75
                    stock_quantity: 99
            example:
              name: boby
              price: 456.75
              stock_quantity: 99
        required: true
      responses:
        "200":
          description: "Product updated successfully."
          headers: {}
        "400":
          description: "Invalid token or product does not exist."
        "401":
          description: "No token."
        "403":
          description: "You don't have the permssions to access this route."
      deprecated: false
      security:
        - bearer: []
  /product:
    post:
      tags:
        - Product
      summary: Create product
      operationId: Createproduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - brand
                - description
                - available
                - stock_quantity
                - price
                - image_url
              properties:
                name:
                  type: string
                  example: "A test name"
                brand:
                  type: string
                  example: "A brand name"
                description:
                  type: string
                  example: "A product description"
                available:
                  type: string
                  enum: [yes, no]
                  example: "yes"
                stock_quantity:
                  type: integer
                  minimum: 0
                  example: 5
                price:
                  type: number
                  format: float
                  example: 9999.99
                image_url:
                  type: array
                  items:
                    type: string
                    format: url
      responses:
        "200":
          description: "Product added to database."
          headers: {}
        "400":
          description: "Invalid token"
        "401":
          description: "No token."
        "403":
          description: "You don't have the permssions to access this route."
        "409":
          description: "Product already exists in the database."
      deprecated: false
      security:
        - bearer: []
...
components:
  schemas:
...
    UpdateproductRequest:
      title: UpdateproductRequest
      required:
        - name
        - price
        - stock_quantity
      type: object
      properties:
        name:
          type: string
        price:
          type: number
        stock_quantity:
          type: integer
          format: int32
      example:
        name: "Egg"
        price: 456.75
        stock_quantity: 99
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
security: []
...